Added a Player (non-functional), Enemy (semi-functional)
Added an object pool script that instantiates a given amount of game objects.
Added an enemy that moves and orients toward the player, and can deal damage to the player and returns to the enemy pool after damaging the player.
Added a enemy spawner that takes enemies from the pool and spawns them off screen with adjustable spawn rate.
And started working on a wave system.